backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
python-dotenv>=1.0.1
shioaji>=1.1.0
duckdb>=1.0.0
yfinance>=0.2.40
google-generativeai>=0.5.4
pandas>=2.2.2
pandas-ta>=0.3.14b0
